Computing Nash Equilibria

Computing Nash Equilibria

This is documentation for Computing Nash Equilibria.

Pure Nash

Documentation for pure_nash.jl.

Exported

Games.pure_nash — Method.

Finds all pure action Nash equilibria for a normal form game. It returns an empty array if there is no pure action Nash.

Currently uses a brute force algorithm, but that hopefully will change in the future.

source

Internal