Game Generators

Game Generators

Exported

Games.covariance_game — Method.

covariance_game([rng=GLOBAL_RNG], nums_actions, rho)

Return a random N-player NormalFormGame instance with N>=2 where the payoff profiles are drawn independently from the standard multi-normal with the covariance of any pair of payoffs equal to rho, as studied in Rinott and Scarsini (2000).

Arguments

rng::AbstractRNG=GLOBAL_RNG: Random number generator used.

nums_actions::NTuple{N,Int}: Tuple of the numbers of actions, one for each player.

rho::Real: Covariance of a pair of payoff values. Must be in [-1/(N-1), 1], where N is the number of players.

Returns

::NormalFormGame: The generated random N-player NormalFormGame.

References

Y. Rinott and M. Scarsini, "On the Number of Pure Strategy Nash Equilibria in Random Games," Games and Economic Behavior (2000), 274-293.

source

Games.random_game — Method.

random_game([rng=GLOBAL_RNG], nums_actions)

Return a random N-player NormalFormGame instance where the payoffs are drawn independently from the uniform distribution on [0, 1).

Arguments

rng::AbstractRNG=GLOBAL_RNG: Random number generator used.

nums_actions::NTuple{N,Int}: Tuple of the numbers of actions, one for each player.

Returns

::NormalFormGame: The generated random N-player NormalFormGame.

source

Internal