Localint

This is documentation for localint.jl.

Exported

GameTheory.AbstractRevision — Type

AbstractRevision

Abstract type representing revision method.

GameTheory.AsynchronousRevision — Type

AsynchronousRevision

Type representing an asynchronous revision.

GameTheory.LocalInteraction — Type

LocalInteraction{N, T, S, A, TR}

Type representing the local interaction model with N players.

Fields

players::NTuple{N,Player{2,T}} : Tuple of Player instances.

num_actions::Integer : The number of actions for players.

adj_matrix::Array{S,2} : Adjacency matrix of the graph in the model.

revision<:AbstractRevision : The way to revise the action profile.

GameTheory.LocalInteraction — Method

LocalInteraction(payoff_matrix,
                 adj_matrix[, revision=SimultaneousRevision()])

Construct a LocalInteraction instance.

Arguments

payoff_matrix::Matrix : The payoff matrix of the game.

adj_matrix::AbstractMatrix : Adjacency matrix of the graph in the model.

revision::AbstractRevision : Arguments to specify the revision method. SimultaneousRevision() or AsynchronousRevision

Returns

::LocalInteraction : The local interaction model.

GameTheory.LocalInteraction — Method

LocalInteraction(g, adj_matrix[, revision=SimultaneousRevision()])

Construct a LocalInteraction instance.

Arguments

g::NormalFormGame : The game used in the model.

adj_matrix::AbstractMatrix : Adjacency matrix of the graph in the model.

revision::AbstractRevision : Arguments to specify the revision method; SimultaneousRevision() or AsynchronousRevision().

Returns

::LocalInteraction : The local interaction model.

GameTheory.SimultaneousRevision — Type

SimultaneousRevision

Type representing a simultaneous revision.

Internal

GameTheory.time_series! — Method

time_series!(li, out, options, player_ind_seq)

Update the matrix out which is used in time_series method given player index sequence.

Arguments

li::LocalInteraction{N} : LocalInteraction instance.

out::Matrix{<:Integer} : Matrix representing a time series of action profiles.

options::BROptions : Options for best_response method.

player_ind_seq::Vector{<:Integer} : Vector representing the index of players to take an action.

Returns

out::Matrix{<:Integer} : Updated out.

GameTheory.time_series! — Method

time_series!(li, out, options)

Update the matrix out which is used in time_series method. All players take their actions simultaneously.

Arguments

li::LocalInteraction{N} : LocalInteraction instance.

out::Matrix{<:Integer} : Matrix representing a time series of action profiles.

options::BROptions : Options for best_response method.

Returns

out::Matrix{<:Integer} : Updated out.