Logitdyn

This is documentation for logitdyn.jl.

Exported

GameTheory.LogitDynamics — Type

LogitDynamics{N, T, S}

Type representing the Logit-Dynamics model.

Fields

players::NTuple{N,Player{N,T}} : Tuple of Player instances.

nums_actions::NTuple{N,Int} : Tuple of the numbers of actions, one for each player.

beta<:Real : The level of noise in a player's decision.

choice_probs::Vector{Array} : The choice probabilities of each action, one for each player.

GameTheory.LogitDynamics — Method

LogitDynamics(g, beta)

Construct a LogitDynamics instance.

Arguments

g::NormalFormGame{N,T} : NormalFormGame instance.

beta::S : The level of noise in players' decision.

Returns

::LogitDynamics : The Logit-Dynamics model.

GameTheory.play! — Method

play!(rng, ld, player_ind, actions)

Return a new action of player indexed by player_ind given each players' choice probabilities.

Arguments

rng::AbstractRNG : Random number generator used.

ld::LogitDynamics{N} : LogitDynamics instance.

player_ind::Integer : A player index who takes an action.

actions::Vector{<:Integer} : The action profile.

Returns

::Integer : The new action of the player indexed by player_ind.

GameTheory.play — Method

play([rng=Random.GLOBAL_RNG,] ld, init_actions[; num_reps=1])

Return new action profile after num_reps iterations.

Arguments

rng::AbstractRNG : Random number generator used.

ld::LogitDynamics{N} : LogitDynamics instance.

init_actions::PureActionProfile : Initial action profile.

num_reps::Integer : The number of iterations.

Returns

::Vector{<:Integer} : New action profile.

GameTheory.time_series — Method

time_series([rng=Random.GLOBAL_RNG,] ld, ts_length, init_actions)

Return a time series of action profiles.

Arguments

rng::AbstractRNG : Random number generator used.

ld::LogitDynamics{N} : LogitDynamics instance.

ts_length::Integer : The length of time series.

init_actions::PureActionProfile : Initial action profile.

Returns

::Matrix{<:Integer} : The time series of action profiles.

Internal

GameTheory.time_series! — Method

time_series!(rng, ld, out, player_ind_seq)

Update the matrix out which is used in time_series method given a player index sequence.

Arguments

rng::AbstractRNG : Random number generator used.

ld::LogitDynamics{N} : LogitDynamics instance.

out::Matrix{<:Integer} : Matrix representing the time series of action profiles.

player_ind_seq::Vector{<:Integer} : The sequence of player index, which is determined randomly.

Returns

::Matrix{<:Integer} : Updated out.