Normal Form Game

This is documentation for normal_form_game.jl.

Exported

Games.NormalFormGame — Type.

Class representing an N-player normal form game.

Fields

players::NTuple{N,Player{N,T<:Real}} : Tuple of Player instances.

N::Int : The number of players.

nums_actions::NTuple{N,Int} : Tuple of the numbers of actions, one for each

player.

source

Games.NormalFormGame — Method.

Constructor of an N-player NormalFormGame.

Arguments

players::Vector{Player} : Vector of Player instances.

source

Games.NormalFormGame — Method.

NormalFormGame{T<:Real}(payoffs::Matrix{T})

Construct a symmetric 2-player NormalFormGame with a square matrix.

Arguments

payoffs::Matrix{T<:Real} : Square matrix representing each player's payoff matrix.

source

Games.NormalFormGame — Method.

NormalFormGame{T<:Real,M}(payoffs::Array{T,M})

Construct an N-player NormalFormGame for N>=2 with an array payoffs of M=N+1 dimensions, where payoffs[a_1, a_2, ..., a_N, :] contains a profile of N payoff values.

Arguments

payoffs::Array{T<:Real} : Array with ndims=N+1 containing payoff profiles.

source

Games.NormalFormGame — Method.

Constructor of an N-player NormalFormGame.

Arguments

players::NTuple{N,Player} : Tuple of Player instances.

source

Games.NormalFormGame — Method.

Constructor of an N-player NormalFormGame, consisting of payoffs all 0.

Arguments

T::Type : Type of payoff values; defaults to Float64 if not specified.

nums_actions::NTuple{N,Int} : Numbers of actions of the N players.

source

Games.NormalFormGame — Method.

Constructor of an N-player NormalFormGame.

Arguments

players::Player{N,T}... : N Player instances

Examples

# p1, p2, and p3 are all of type `Player{3,T}` for some `T`
NormalFormGame(p1, p2, p3)

source

Games.Player — Type.

Type representing a player in an N-player normal form game.

Arguments

payoff_array::Array{T<:Real} : Array representing the player's payoff

function.

Fields

payoff_array::Array{T<:Real} : Array representing the player's payoff

function.

source

Games.best_response — Method.

Return the perturbed best response to opponents_actions.

Arguments

player::Player : Player instance.

opponents_actions::Union{Action,ActionProfile,Void} : Profile of N-1

opponents' actions. If N=2, then it must be a vector of reals (in which case it is treated as the opponent's mixed action) or a scalar of integer (in which case it is treated as the opponent's pure action). If N>2, then it must be a tuple of N-1 integers (pure actions) or N-1 vectors of reals (mixed actions). (For the degenerate case N=1, it must be nothing.)

payoff_perturbation::Vector{Float64} : Vector of length equal to the number

of actions of the player containing the values ("noises") to be added to the payoffs in determining the best response.

Returns

::Int : Best response action.

source

Games.best_response — Method.

Return a best response action to opponents_actions.

Arguments

player::Player : Player instance.

opponents_actions::Union{Action,ActionProfile,Void} : Profile of N-1

opponents' actions. If N=2, then it must be a vector of reals (in which case it is treated as the opponent's mixed action) or a scalar of integer (in which case it is treated as the opponent's pure action). If N>2, then it must be a tuple of N-1 integers (pure actions) or N-1 vectors of reals (mixed actions). (For the degenerate case N=1, it must be nothing.)

tie_breaking::AbstractString("smallest") : Control how to break a tie (see

Returns for details).

tol::Float64 : Tolerance to be used to determine best response actions.

Returns

::Int : If tie_breaking="smallest", returns the best response action with

the smallest index; if tie_breaking="random", returns an action randomly chosen from the best response actions.

source

Games.best_responses — Method.

Return all the best response actions to opponents_actions.

Arguments

player::Player : Player instance.

opponents_actions::Union{Action,ActionProfile,Void} : Profile of N-1

opponents' actions. If N=2, then it must be a vector of reals (in which case it is treated as the opponent's mixed action) or a scalar of integer (in which case it is treated as the opponent's pure action). If N>2, then it must be a tuple of N-1 integers (pure actions) or N-1 vectors of reals (mixed actions). (For the degenerate case N=1, it must be nothing.)

;tol::Float64 : Tolerance to be used to determine best response actions.

Returns

best_responses::Vector{Int} : Vector containing all the best response

actions.

source

Games.is_best_response — Method.

Return true if own_action is a best response to opponents_actions.

Arguments

player::Player : Player instance.

own_action::MixedAction : Own mixed action (vector of reals).

opponents_actions::Union{Action,ActionProfile,Void} : Profile of N-1

opponents' actions. If N=2, then it must be a vector of reals (in which case it is treated as the opponent's mixed action) or a scalar of integer (in which case it is treated as the opponent's pure action). If N>2, then it must be a tuple of N-1 integers (pure actions) or N-1 vectors of reals (mixed actions). (For the degenerate case N=1, it must be nothing.)

;tol::Float64 : Tolerance to be used to determine best response actions.

Returns

::Bool : True if own_action is a best response to opponents_actions;

false otherwise.

source

Games.is_best_response — Method.

Return True if own_action is a best response to opponents_actions.

Arguments

player::Player : Player instance.

own_action::PureAction : Own pure action (integer).

opponents_actions::Union{Action,ActionProfile,Void} : Profile of N-1

opponents' actions. If N=2, then it must be a vector of reals (in which case it is treated as the opponent's mixed action) or a scalar of integer (in which case it is treated as the opponent's pure action). If N>2, then it must be a tuple of N-1 integers (pure actions) or N-1 vectors of reals (mixed actions). (For the degenerate case N=1, it must be nothing.)

;tol::Float64 : Tolerance to be used to determine best response actions.

Returns

::Bool : True if own_action is a best response to opponents_actions;

valse otherwise.

source

Games.is_nash — Function.

Return true if action_profile is a Nash equilibrium.

Arguments

g::NormalFormGame : Instance of N-player NormalFormGame.

action_profile::ActionProfile : Tuple of N integers (pure actions) or N

vectors of reals (mixed actions).

Returns

::Bool

source

Games.is_nash — Method.

Return true if action is a Nash equilibrium of a trivial game with 1 player.

Arguments

g::NormalFormGame : Instance of 1-player NormalFormGame.

action::Action : Integer (pure action) or vector of reals (mixed action).

Returns

::Bool

source

Games.payoff_vector — Method.

Return a vector of payoff values for a Player in an N>2 player game, one for each own action, given a tuple of the opponents' pure actions.

Arguments

player::Player : Player instance.

opponents_actions::PureActionProfile : Tuple of N-1 opponents' pure

actions.

Returns

::Vector : Payoff vector.

source

Games.payoff_vector — Method.

Return a vector of payoff values for a Player in a trivial game with 1 player, one for each own action.

Arguments

player::Player : Player instance.

opponent_action::Void

Returns

::Vector : Payoff vector.

source

Games.payoff_vector — Method.

Return a vector of payoff values for a Player in a 2-player game, one for each own action, given the opponent's mixed action.

Arguments

player::Player : Player instance.

opponent_action::MixedAction : Opponent's mixed action (vector of reals).

Returns

::Vector : Payoff vector.

source

Games.payoff_vector — Method.

Return a vector of payoff values for a Player in a 2-player game, one for each own action, given the opponent's pure action.

Arguments

player::Player : Player instance.

opponent_action::PureAction : Opponent's pure action (integer).

Returns

::Vector : Payoff vector.

source

Games.payoff_vector — Method.

Return a vector of payoff values for a Player in an N>2 player game, one for each own action, given a tuple of the opponents' mixed actions.

Arguments

player::Player : Player instance.

opponents_actions::MixedActionProfile : Tuple of N-1 opponents' mixed

actions.

Returns

::Vector : Payoff vector.

source

Games.pure2mixed — Method.

Convert a pure action to the corresponding mixed action.

Arguments

num_actions::Integer : The number of the pure actions (= the length of a

mixed action).

action::PureAction : The pure action to convert to the corresponding mixed

action.

Returns

mixed_action::Vector{Float64} : The mixed action representation of the

given pure action.

source

Internal