Repeated Game

Repeated Game

This is documentation for Repeated Game.

Repeated Game Utilities

Documentation for repeated_game_util.jl.

Exported

Internal

Games.unitcircle — Method.

Places npts equally spaced points along the 2 dimensional circle and returns the points with x coordinates in first column and y coordinates in second column

i.e. if you wanted point i, it would be pts[i, :]

source

Repeated Game

Documentation for repeated_game.jl.

Exported

Games.RepeatedGame — Type.

This is a type for a specific type of repeated games

It takes a stage game that is repeated in every period and all agents discount future at rate δ

source

Games.RepeatedGame — Method.

Helper constructor that builds game from players

source

Games.outerapproximation — Method.

Approximates the set of equilibrium value set for a repeated game with the outer hyperplane approximation described by Judd, Yeltekin, Conklin 2002

NOTE: If your code fails then it might be the case that the value set is only the value which corresponds to the pure action nash equilibrium or you might just need more points to be precise enough for the Polyhedra library to be able to convert to the vertice representation.

The arguments are

rpd: 2 player repeated game

The keyword arguments are

nH: Number of subgradients used in approximation

tol: Tolerance in differences of set

maxiter: Maximum number of iterations

verbose: Whether to display updates about iterations and distance

nskipprint: Number of iterations between printing information (verbose=true)

check_pure_nash: Whether to perform a check about whether a pure Nash equilibrium exists

plib: Allows users to choose a particular package for the geometry computations (See Polyhedra.jl docs for more info). By default, it chooses to use CDDLib.jl

source

Games.unpack — Method.

Unpacks the elements of a repeated game

source

Games.worst_value_1 — Method.

See worst_value_i for documentation

source

Games.worst_value_2 — Method.

See worst_value_i for documentation

source

Games.worst_value_i — Method.

Given a constraint w ∈ W, this finds the worst possible payoff for agent i

The output of this function is used to create the values associated with incentive compatibility constraints

The arguments for this function are

rpd: Two player repeated game

H: Subgradients used to approximate value set

C: Hyperplane levels for value set approximation

i: Which player want worst value for

source

Games.worst_values — Method.

See worst_value_i for documentation

source

Internal

Games.initialize_LP_matrices — Method.

Initialize matrices for the linear programming problems. It sets up the A matrix since it never changes, but only allocates space for b and c since they will be filled repeatedly on the iterations

We add nH slack variables (which will be constrained to be positive) to deal with inequalities associated with Ax leq b.

min c ⋅ x Ax < b

In this case, the c vector will be determined by which subgradient is being used, so this function only allocates space for it.

The A matrix will be filled with nH set constraints and 2 incentive compatibility constraints. The set constraints restrain the linear programming problem to pick solutions that are in the current set of continuation values while the incentive compatibility constraints ensure the agents won't deviate.

The b vector is associated with the A matrix and gives the value for constraint.

The arguments for this function are

rpd: Two player repeated game

H: The subgradients used to approximate the value set

source

Games.initialize_sg_hpl — Method.

This is a function that initializes the subgradients, hyperplane levels, and extreme points of the value set by choosing an appropriate origin and radius.

See initialize_sg_hpl for more documentation

source

Games.initialize_sg_hpl — Method.

Initializes the following things for a 2 player repeated game.

subgradients

extreme points of the convex set for values

hyper plane levels

These are determined in the following way

Subgradients are simply chosen from the unit circle.

The values for the extremum of the value set are just given by choosing points along a circle with specified origin and radius

Hyperplane levels are determined by computing the hyperplane level such that the extreme points from the circle are generated

The arguments are

source