Repeated Games

Repeated Games

Exported

Games.RepeatedGame — Type.

RepeatedGame{N,T}

Class representing an N-player repeated game.

Fields

sg::NormalFormGame{N, T} : The stage game used to create the repeated game.

delta::Float64 : The common discount rate at which all players discount the future.

source

Games.RepeatedGame — Method.

RepeatedGame(p1, p2, delta)

Helper constructor that builds a repeated game for two players.

Arguments

p1::Player : The first player.

p2::Player : The second player.

delta::Float64 : The common discount rate at which all players discount the future.

Returns

::RepeatedGame : The repeated game.

source

Games.outerapproximation — Method.

outerapproximation(rpd; nH=32, tol=1e-8, maxiter=500, check_pure_nash=true,
                   verbose=false, nskipprint=50,
                   plib=getlibraryfor(2, Float64))

Approximates the set of equilibrium value set for a repeated game with the outer hyperplane approximation described by Judd, Yeltekin, Conklin 2002.

Arguments

rpd::RepGame2 : Two player repeated game.

nH : Number of subgradients used for the approximation.

tol : Tolerance in differences of set.

maxiter : Maximum number of iterations.

verbose : Whether to display updates about iterations and distance.

nskipprint : Number of iterations between printing information (assuming verbose=true).

check_pure_nash: Whether to perform a check about whether a pure Nash equilibrium exists.

plib: Allows users to choose a particular package for the geometry computations. (See Polyhedra.jl docs for more info). By default, it chooses to use SimplePolyhedraLibrary.

Returns

vertices::Array{Float64} : Vertices of the outer approximation of the value set.

source

Games.unpack — Method.

unpack(rpd)

Helper function that unpacks the elements of a repeated game.

source

Games.worst_value_1 — Method.

See worst_value_i for documentation

source

Games.worst_value_2 — Method.

See worst_value_i for documentation

source

Games.worst_value_i — Method.

worst_value_i(rpd, H, C, i)

Given a constraint w ∈ W, this finds the worst possible payoff for agent i.

Arugments

rpd::RepGame2 : Two player repeated game.

H::Array{Float64, 2} : The subgradients used to approximate the value set.

C::Array{Float64, 1} : The array containing the hyperplane levels.

i::Int : The player of interest.

Returns

out::Float64 : Worst possible payoff for player i.

source

Games.worst_values — Method.

See worst_value_i for documentation

source

Internal

Games.initialize_LP_matrices — Method.

initialize_LP_matrices(rpd, H)

Initialize matrices for the linear programming problems.

Arguments

rpd::RepeatedGame : Two player repeated game.

H : The subgradients used to approximate the value set.

Returns

c::Array{Float64}(nH, 1) : Vector used to determine which subgradient is being used.

A::Array{Float64}(nH, 2) : Matrix with nH set constraints and to be filled with 2 additional incentive compatibility constraints.

b::Array{Float64}(nH, 1) : Vector to be filled with the values for the constraints.

source

Games.initialize_sg_hpl — Method.

initialize_sg_hpl(rpd, nH)

Initializes subgradients, extreme points and hyperplane levels for the approximation of the convex value set of a 2 player repeated game by choosing an appropriate origin and radius.

Arguments

rpd::RepeatedGame : Two player repeated game.

nH::Int : Number of subgradients used for the approximation.

Returns

C::Array{Float64}(nH, 1) : The array containing the hyperplane levels.

H::Array{Float64}(nH, 2) : The array containing the subgradients.

Z::Array{Float64}(nH, 2) : The array containing the extreme points of the value set.

source

Games.initialize_sg_hpl — Method.

initialize_sg_hpl(nH, o, r)

Initializes subgradients, extreme points and hyperplane levels for the approximation of the convex value set of a 2 player repeated game.

Arguments

nH::Int : Number of subgradients used for the approximation.

o::Vector{Float64} : Origin for the approximation.

r::Float64 : Radius for the approximation.

Returns

C::Array{Float64}(nH, 1) : The array containing the hyperplane levels.

H::Array{Float64}(nH, 2) : The array containing the subgradients.

Z::Array{Float64}(nH, 2) : The array containing the extreme points of the value set.

source

Games.unitcircle — Method.

unitcircle(npts)

Places npts equally spaced points along the 2 dimensional circle and returns the points with x coordinates in first column and y coordinates in second column.

source